+++ src/managers/InfoManager.ts
@@ -1,7 +1,8 @@
 import fetch from "isomorphic-fetch"
 import { ManagerOptions } from "../types/types"
-import { Webhook } from "../types/apitypes"
+import { Webhook } from "fagc-api-types"
 import BaseManager from "./BaseManager"
+import { GenericAPIError } from "../types"
 
 export default class InfoManager extends BaseManager<Webhook> {
 	public apikey?: string
@@ -21,6 +22,8 @@ export default class InfoManager extends BaseManager<Webhook> {
 			}),
 			headers: { "content-type": "application/json" },
 		}).then(w=>w.json())
+		console.log({add})
+		if (add.error) throw new GenericAPIError(`${add.error}: ${add.description}`)
 		return add
 	} 
 	async removeWebhook(webhookid: string, webhooktoken: string, guildid: string): Promise<Webhook> {