+++ src/types/index.ts
@@ -1,3 +1,2 @@
-export * from "./apitypes"
 export * from "./types"
 export * from "./errors"